+++ app/models/lexeme.rb
@@ -85,7 +85,7 @@ class Lexeme < ActiveRecord::Base
     when SUBSTRING
       headwords_like = "(headword_translations.form LIKE ?" + " OR headword_translations.form LIKE ?" * (forms.length - 1) + ")"
       wildcarded_forms = forms.collect {|form| "%#{form}%"}
-      Lexeme.joins(:headwords => [:translations]).where([headwords_like, *wildcarded_forms]).includes(options[:include]).group(:lexeme_id)
+      Lexeme.joins(:headwords => [:translations]).where([headwords_like, *wildcarded_forms]).includes(options[:include]).group('"headwords"."lexeme_id"')
     when EXACT
       Lexeme.joins(:headwords => [:translations]).where(["headword_translations.form IN (?)", forms]).includes(options[:include])
     end